+++ github/includes/class-github-updater.php
@@ -54,6 +54,12 @@ class GitHub_Updater {
 	protected static $extra_headers = array();
 
 	/**
+	 * Holds the values to be used in the fields callbacks
+	 * @var array
+	 */
+	protected static $options;
+
+	/**
 	 * Autoloader
 	 *
 	 * @param $class
@@ -78,17 +84,15 @@ class GitHub_Updater {
 	/**
 	 * Constructor
 	 *
-	 * Calls $this->init() in init hook so other remote upgrader apps like
-	 * InfiniteWP, ManageWP, MainWP, and iThemes Sync will load and use all
-	 * of GitHub_Updater's methods, especially renaming.
-	 *
 	 * Calls spl_autoload_register to set loading of classes.
+	 * Loads options to private static variable.
 	 */
 	public function __construct() {
-		//add_action( 'init', array( $this, 'init' ) );
 		if ( function_exists( 'spl_autoload_register' ) ) {
 			spl_autoload_register( array( $this, 'autoload' ) );
 		}
+		self::$options = get_site_option( 'github_updater' );
+		$this->add_headers();
 	}
 
 	/**
@@ -105,6 +109,58 @@ class GitHub_Updater {
 		if ( is_admin() && ( current_user_can( 'update_plugins' ) || current_user_can( 'update_themes' ) ) ) {
 			new GitHub_Updater_Settings;
 		}
+	}
+
+	/**
+	 * Add extra headers via filter hooks
+	 */
+	public static function add_headers() {
+		add_filter( 'extra_plugin_headers', array( __CLASS__, 'add_plugin_headers' ) );
+		add_filter( 'extra_theme_headers', array( __CLASS__, 'add_theme_headers' ) );
+	}
+
+	/**
+	 * Add extra headers to get_plugins();
+	 *
+	 * @param $extra_headers
+	 * @return array
+	 */
+	public static function add_plugin_headers( $extra_headers ) {
+		$ghu_extra_headers   = array(
+			'GitHub Plugin URI'    => 'GitHub Plugin URI',
+			'GitHub Branch'        => 'GitHub Branch',
+			'GitHub Access Token'  => 'GitHub Access Token',
+			'Bitbucket Plugin URI' => 'Bitbucket Plugin URI',
+			'Bitbucket Branch'     => 'Bitbucket Branch',
+			'Requires WP'          => 'Requires WP',
+			'Requires PHP'         => 'Requires PHP',
+		);
+		self::$extra_headers = array_unique( array_merge( self::$extra_headers, $ghu_extra_headers ) );
+		$extra_headers       = array_merge( (array) $extra_headers, (array) $ghu_extra_headers );
+
+		return $extra_headers;
+	}
+
+	/**
+	 * Add extra headers to wp_get_themes()
+	 *
+	 * @param $extra_headers
+	 * @return array
+	 */
+	public static function add_theme_headers( $extra_headers ) {
+		$ghu_extra_headers   = array(
+			'GitHub Theme URI'    => 'GitHub Theme URI',
+			'GitHub Branch'       => 'GitHub Branch',
+			'GitHub Access Token' => 'GitHub Access Token',
+			'Bitbucket Theme URI' => 'Bitbucket Theme URI',
+			'Bitbucket Branch'    => 'Bitbucket Branch',
+			'Requires WP'         => 'Requires WP',
+			'Requires PHP'        => 'Requires PHP',
+		);
+		self::$extra_headers = array_unique( array_merge( self::$extra_headers, $ghu_extra_headers ) );
+		$extra_headers       = array_merge( (array) $extra_headers, (array) $ghu_extra_headers );
+
+		return $extra_headers;
 	}
 
 	/**
@@ -153,7 +209,6 @@ class GitHub_Updater {
 	*/
 	protected function get_local_plugin_meta( $headers ) {
 		$git_repo = array();
-		$options  = get_site_option( 'github_updater' );
 
 		// Reverse sort to run plugin/theme URI first
 		arsort( self::$extra_headers );
@@ -187,9 +242,9 @@ class GitHub_Updater {
 					if ( empty( $headers['GitHub Access Token'] ) ) {
 						break;
 					}
-					$git_repo['access_token']  = $headers['GitHub Access Token'];
+					$git_repo['access_token'] = $headers['GitHub Access Token'];
 
-					$this->save_header_options( $git_repo['repo'], $git_repo['access_token'], $options );
+					$this->save_header_options( $git_repo['repo'], $git_repo['access_token'], self::$options );
 					break;
 			}
 		}
@@ -205,8 +260,6 @@ class GitHub_Updater {
 					}
 					$git_repo['type']       = 'bitbucket_plugin';
 
-					$git_repo['user']       = parse_url( $headers['Bitbucket Plugin URI'], PHP_URL_USER );
-					$git_repo['pass']       = parse_url( $headers['Bitbucket Plugin URI'], PHP_URL_PASS );
 					$owner_repo             = parse_url( $headers['Bitbucket Plugin URI'], PHP_URL_PATH );
 					$owner_repo             = trim( $owner_repo, '/' );  // strip surrounding slashes
 					$git_repo['uri']        = 'https://bitbucket.org/' . $owner_repo;
@@ -214,8 +267,6 @@ class GitHub_Updater {
 					$git_repo['owner']      = $owner_repo[0];
 					$git_repo['repo']       = $owner_repo[1];
 					$git_repo['local_path'] = WP_PLUGIN_DIR . '/' . $git_repo['repo'] .'/';
-
-					$this->save_header_options( $git_repo['repo'], $git_repo['pass'], $options );
 					break;
 				case 'Bitbucket Branch':
 					if ( empty( $headers['Bitbucket Branch'] ) ) {
@@ -230,40 +281,15 @@ class GitHub_Updater {
 	}
 
 	/**
-	* Get array of all themes in multisite
-	*
-	* wp_get_themes does not seem to work under network activation in the same way as in a single install.
-	* http://core.trac.wordpress.org/changeset/20152
-	*
-	* @return array
-	*/
-	protected function multisite_get_themes() {
-		$themes     = array();
-		$theme_dirs = scandir( get_theme_root() );
-		$theme_dirs = array_diff( $theme_dirs, array( '.', '..', '.DS_Store', 'index.php' ) );
-
-		foreach ( (array) $theme_dirs as $theme_dir ) {
-			$themes[] = wp_get_theme( $theme_dir );
-		}
-
-		return $themes;
-	}
-
-	/**
 	 * Reads in WP_Theme class of each theme.
 	 * Populates variable array
 	 */
 	protected function get_theme_meta() {
 		$git_themes = array();
-		$themes     = wp_get_themes();
-		$options    = get_site_option( 'github_updater' );
+		$themes     = wp_get_themes( array( 'errors' => null ) );
 
 		// Reverse sort to run plugin/theme URI first
 		arsort( self::$extra_headers );
-
-		if ( is_multisite() ) {
-			$themes = $this->multisite_get_themes();
-		}
 
 		foreach ( (array) $themes as $theme ) {
 			$git_theme         = array();
@@ -287,7 +313,6 @@ class GitHub_Updater {
 						if ( empty( $github_uri ) ) {
 							break;
 						}
-
 						$git_theme['type']                    = 'github_theme';
 
 						$owner_repo                           = parse_url( $github_uri, PHP_URL_PATH );
@@ -315,7 +340,7 @@ class GitHub_Updater {
 						}
 						$git_theme['access_token']            = $github_token;
 
-						$this->save_header_options( $git_theme['repo'], $github_token, $options );
+						$this->save_header_options( $git_theme['repo'], $github_token, self::$options );
 						break;
 				}
 			}
@@ -329,11 +354,8 @@ class GitHub_Updater {
 						if ( empty( $bitbucket_uri ) ) {
 							break;
 						}
-
 						$git_theme['type']                    = 'bitbucket_theme';
 
-						$git_theme['user']                    = parse_url( $bitbucket_uri, PHP_URL_USER );
-						$git_theme['pass']                    = parse_url( $bitbucket_uri, PHP_URL_PASS );
 						$owner_repo                           = parse_url( $bitbucket_uri, PHP_URL_PATH );
 						$owner_repo                           = trim( $owner_repo, '/' );
 						$git_theme['uri']                     = 'https://bitbucket.org/' . $owner_repo;
@@ -346,8 +368,6 @@ class GitHub_Updater {
 						$git_theme['local_version']           = $theme->get( 'Version' );
 						$git_theme['sections']['description'] = $theme->get( 'Description' );
 						$git_theme['local_path']              = get_theme_root() . '/' . $git_theme['repo'] .'/';
-
-						$this->save_header_options( $git_theme['repo'], $git_theme['pass'], $options );
 						break;
 					case 'Bitbucket Branch':
 						if ( empty( $bitbucket_branch ) ) {
@@ -370,10 +390,9 @@ class GitHub_Updater {
 	 * @param $type
 	 */
 	protected function set_defaults( $type ) {
-		$options = get_site_option( 'github_updater' );
-		if ( ! isset( $options[ $this->$type->repo ] ) ) {
-			$options[ $this->$type->repo ] = null;
-			add_site_option( 'github_updater', $options );
+		if ( ! isset( self::$options[ $this->$type->repo ] ) ) {
+			self::$options[ $this->$type->repo ] = null;
+			add_site_option( 'github_updater', self::$options );
 		}
 
 		$this->$type->remote_version        = '0.0.0';
@@ -390,6 +409,7 @@ class GitHub_Updater {
 		$this->$type->num_ratings           = 0;
 		$this->$type->transient             = array();
 		$this->$type->repo_meta             = array();
+		$this->$type->private               = true;
 		$this->$type->watchers              = 0;
 		$this->$type->forks                 = 0;
 		$this->$type->open_issues           = 0;
@@ -401,9 +421,9 @@ class GitHub_Updater {
 	/**
 	 * Rename the zip folder to be the same as the existing repository folder.
 	 *
-	 * Github delivers zip files as <Repo>-<Branch>.zip
+	 * Github delivers zip files as <User>-<Repo>-<Branch|Hash>.zip
 	 *
-	 * @global WP_Filesystem $wp_filesystem
+	 * @global object $wp_filesystem
 	 *
 	 * @param string $source
 	 * @param string $remote_source Optional.
@@ -593,13 +613,14 @@ class GitHub_Updater {
 	 */
 	protected function delete_all_transients( $type ) {
 		$transients = get_site_transient( 'ghu-' . $type );
-		if ( empty( $transients ) ) {
+		if ( ! $transients ) {
 			return false;
 		}
 
 		foreach ( $transients as $transient ) {
 			delete_site_transient( $transient );
 		}
+		delete_site_transient( 'ghu-' . $type );
 	}
 
 
@@ -607,9 +628,13 @@ class GitHub_Updater {
 	 * Create transient of $type transients for force-check
 	 *
 	 * @param $type
-	 * @return void
+	 * @return void|bool
 	 */
 	protected function make_force_check_transient( $type ) {
+		$transient = get_site_transient( 'ghu-' . $type );
+		if ( $transient ) {
+			return false;
+		}
 		set_site_transient( 'ghu-' . $type , self::$transients, self::$hours * HOUR_IN_SECONDS );
 		self::$transients = array();
 	}
@@ -624,10 +649,10 @@ class GitHub_Updater {
 	 * @return float|int
 	 */
 	protected function make_rating( $repo_meta ) {
-		$watchers    = ( empty( $repo_meta->watchers ) ? $this->type->watchers : $repo_meta->watchers );
-		$forks       = ( empty( $repo_meta->forks ) ? $this->type->forks : $repo_meta->forks );
-		$open_issues = ( empty( $repo_meta->open_issues ) ? $this->type->open_issues : $repo_meta->open_issues );
-		$score       = ( empty( $repo_meta->score ) ? $this->type->score : $repo_meta->score ); //what is this anyway?
+		$watchers    = empty( $repo_meta->watchers ) ? $this->type->watchers : $repo_meta->watchers;
+		$forks       = empty( $repo_meta->forks ) ? $this->type->forks : $repo_meta->forks;
+		$open_issues = empty( $repo_meta->open_issues ) ? $this->type->open_issues : $repo_meta->open_issues;
+		$score       = empty( $repo_meta->score ) ? $this->type->score : $repo_meta->score; //what is this anyway?
 
 		$rating = round( $watchers + ( $forks * 1.5 ) - $open_issues + $score );
 
@@ -650,8 +675,12 @@ class GitHub_Updater {
 		if ( ! $value ) {
 			return false;
 		}
-		$options[ $repo ] = $value;
-		update_site_option( 'github_updater', $options );
+		if ( empty( $options[ $repo ] )  && ! empty( $value ) ) {
+			unset( $options[ $repo ] );
+			$value = sanitize_text_field( $value );
+			$options[ $repo ] = $value;
+			update_site_option( 'github_updater', $options );
+		}
 	}
 
 	/**